refactor(types): migrate type imports from interfaces to types module
- Updated type imports across multiple files to reference the new 'types' module instead of 'interfaces', improving consistency and organization.
- Introduced new type definitions for actions, analysis configurations, and game structures, enhancing clarity and maintainability of the codebase.
- Streamlined the index file to re-export all types, ensuring existing import paths remain functional while promoting better type management.
