+++ src/types/analysis.ts
